requirements.txt
pyrogram==2.0.106
py-tgcalls==2.0.9
ntgcalls==1.2.7
yt-dlp==2024.12.6
pydantic==2.10.4
pydantic-settings==2.7.1
motor==3.6.0
redis==5.2.1
aiofiles==24.1.0
Pillow==11.0.0
spotipy==2.24.0
aiohttp==3.11.11
apscheduler==3.11.0
python-dotenv==1.0.1
TgCrypto==1.2.5
